+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Fixed
+- issue where prompt=true was not respecting the default values 
+- display of option values was munging booleans
+
+### Added
+- validation of prompted valued for click to ensure non-empty strings
+
 ## 3.0.2
 ### Changed
 - `TILLER_NAMESPACE` is now set to `tiller` by default